requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/dev.in
#
appdirs==1.4.4
    # via black
asgiref==3.3.1
    # via django
backcall==0.2.0
    # via ipython
git+git://github.com/psf/black
    # via -r requirements/dev.in
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   black
    #   pip-tools
decorator==4.4.2
    # via ipython
dj-database-url==0.5.0
    # via django-heroku
django-debug-toolbar==3.2
    # via -r requirements/dev.in
django-environ==0.4.5
    # via -r requirements/base.in
django-extensions==3.1.0
    # via -r requirements/dev.in
django-health-check==3.16.2
    # via -r requirements/base.in
django-heroku==0.3.1
    # via -r requirements/base.in
django==3.1.5
    # via
    #   -r requirements/base.in
    #   django-debug-toolbar
    #   django-health-check
    #   django-heroku
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements/base.in
gunicorn==20.0.4
    # via -r requirements/base.in
idna==2.10
    # via requests
ipython-genutils==0.2.0
    # via traitlets
ipython==7.19.0
    # via -r requirements/dev.in
jedi==0.18.0
    # via ipython
markdown==3.3.3
    # via -r requirements/base.in
mypy-extensions==0.4.3
    # via black
parso==0.8.1
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.1.0
    # via -r requirements/base.in
pip-tools==5.5.0
    # via -r requirements/dev.in
prompt-toolkit==3.0.14
    # via ipython
psycopg2==2.8.6
    # via
    #   -r requirements/base.in
    #   django-heroku
ptyprocess==0.7.0
    # via pexpect
pygments==2.7.4
    # via ipython
python-dateutil==2.8.1
    # via -r requirements/base.in
pytz==2020.5
    # via django
regex==2020.11.13
    # via black
requests==2.25.1
    # via stripe
sentry-sdk==0.19.5
    # via -r requirements/base.in
six==1.15.0
    # via python-dateutil
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
stripe==2.55.1
    # via -r requirements/base.in
toml==0.10.2
    # via black
traitlets==5.0.5
    # via ipython
typed-ast==1.4.2
    # via black
urllib3==1.26.3
    # via
    #   requests
    #   sentry-sdk
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==5.2.0
    # via django-heroku

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
